refactor: Moved utils for making mock datasets into `test_utils.ts`
(cherry picked from commit 3585d7fdb1276796437c314fdf94294525c82645)
